+++ inc/customizer/controls.php
@@ -14,7 +14,6 @@
 if ( ! function_exists( 'storefront_customize_register' ) ) {
 	function storefront_customize_register( $wp_customize ) {
 		$wp_customize->get_setting( 'blogname' )->transport         = 'postMessage';
-		$wp_customize->get_setting( 'blogdescription' )->transport  = 'postMessage';
 		$wp_customize->get_setting( 'header_textcolor' )->transport = 'postMessage';
 
 		// Move background color setting alongside background image